start to break apart the compiler
